requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.1.2
langgraph>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
pypdf>=5.0.0
python-dotenv>=1.0.0
